edit services-> properties now populates correctly.
this one took some SHENANIGANS. Hear me out on this.

To keep the UI simple, we chose to use a select in multiple mode to handle
selecting properties at which a service is available. Because Ant-React uses a
JS Array to store the select `value`. Clojure maps get transmogrified into JS
objects, and are subject to the same rules of equality (i.e., `{:a 1}` does not
equal `{:a 1}`

tl;dr - i made it so we're just using vectors of property ids, then relying on
the properties subscription to get the name of the property for user-friendly
rendering.